Moved handleFileInjections to util.go
